+++ src/fisca/calculettes/base/base.js
@@ -1,6 +1,14 @@
-var BaseCalculette = (function () {
+var __extends = (this && this.__extends) || function (d, b) {
+    for (var p in b) if (b.hasOwnProperty(p)) d[p] = b[p];
+    function __() { this.constructor = d; }
+    d.prototype = b === null ? Object.create(b) : (__.prototype = b.prototype, new __());
+};
+var base_1 = require('../../../base/base');
+var BaseCalculette = (function (_super) {
+    __extends(BaseCalculette, _super);
     function BaseCalculette(params) {
+        _super.call(this, params);
     }
     return BaseCalculette;
-})();
+})(base_1.default.BaseModel);
 exports.BaseCalculette = BaseCalculette;